8/2 deal with manager

+++ src/main/java/com/example/Fproject/database/serviceImpl/DatabaseServiceImpl.java
@@ -19,26 +19,17 @@ public class DatabaseServiceImpl implements DatabaseService {
     @Autowired
     private DeviceService deviceService;
     @Autowired
-    private userdeviceRepository userdeviceRepository;
-    @Autowired
     private managerRepository managerRepository;
+
 
     @Override
     public boolean authorization(String userId, String deviceId) {
-        //user有無操作device權限
-        /*List<UserDevice> users = new ArrayList<>();
-        users = userdeviceRepository.findAll();
-        boolean flag = false;
-        for(UserDevice userDevice : users){
-            if(userDevice.getUser().equals(userId)&&userDevice.getDevice().equals(deviceId)){
-                flag = true;
-                break;
+        User user = userRepository.findById(userId).orElse(null);
+        Set<Device> devices = user.getDevice();
+        for (Device device : devices) {
+            if (device.getDeviceId().equals(deviceId)) {
+                return true;
             }
-        }
-        return flag;*/
-        UserDevice userDevice = userdeviceRepository.findById(userId).orElse(null);
-        if(userDevice.getDevice().getDeviceId().equals(deviceId)){
-            return true;
         }
         return false;
     }
@@ -54,30 +45,25 @@ public class DatabaseServiceImpl implements DatabaseService {
 
     @Override
     public boolean addDevice(String url, String type, String pin, String userId) {
-        //新增device
         String deviceId = generateRandomId();
+        User user = userRepository.findById(userId).orElse(null);
+        Set<Device> devices = user.getDevice();
         Device device = new Device();
         device.setDeviceId(deviceId);
         device.setType(type);
         device.setPin(pin);
-        //使用新增它的user
-        User user = userRepository.findById(userId).orElse(null);
-        //新增userdevice
-        UserDevice userDevice = new UserDevice();
-        userDevice.setDevice(device);
-        userDevice.setUser(user);
-        //新增manager
+        devices.add(device);
+        user.setDevice(devices);
+
+        Set<Manager> managers = device.getManager();
         Manager manager = new Manager();
         manager.setManagerId(userId);
         manager.setDevice(device);
-
+        managers.add(manager);
+        device.setManager(managers);//device 儲存 manager
         try {
-            //儲存userdevice
-            userdeviceRepository.save(userDevice);
-            //儲存device
+            userRepository.save(user);
             deviceRepository.save(device);
-            //儲存manager
-            managerRepository.save(manager);
             return true;
 
         } catch (Exception e) {
@@ -88,30 +74,27 @@ public class DatabaseServiceImpl implements DatabaseService {
 
     }
 
-    /*----------------------------------先他媽別動------------------------------------*/
     @Override
     public boolean alterDevice(String userId, String deviceId, String url) {
-        /*Device device = deviceRepository.findById(deviceId).orElse(null);
-       // Set<User> user = device.getUser();
+        Device device = deviceRepository.findById(deviceId).orElse(null);
+        Set<User> user = device.getUser();
         if (deviceService.isManager(userId, deviceId)) {
-            //device.setUrl(url);
-            //device.setUser(user);
+            device.setUser(user);
             deviceRepository.save(device);
             return true;
-        }*/
+        }
         return false;
     }
-    /*------------------------------------------------------------------------------*/
+
     @Override
     public boolean deleteDevice(String userId, String deviceId) {
         Device device = deviceRepository.findById(deviceId).orElse(null);
-        try {
+        if (deviceService.isManager(userId, deviceId)) {
+            device.setUser(null);
             deviceRepository.delete(device);
             return true;
-        }catch (Exception e){
-            System.out.println("Error" + e.getMessage());
-            return false;
         }
+        return false;
     }
 
     @Override
@@ -132,6 +115,7 @@ public class DatabaseServiceImpl implements DatabaseService {
     public boolean deleteUser(String userId) {
         User user = userRepository.findById(userId).orElse(null);
         try {
+            user.setDevice(null);
             userRepository.delete(user);
             return true;
         }catch(Exception e){
@@ -139,25 +123,23 @@ public class DatabaseServiceImpl implements DatabaseService {
             return false;
         }
     }
-    /*----------------------------------先他媽別動------------------------------------*/
+
     public String getUrl(String deviceId){
-        /*Device device=deviceRepository.findById(deviceId).orElse(null);
+        Device device=deviceRepository.findById(deviceId).orElse(null);
         if(device!=null){
-            return device.getUrl();
+            return null;
         }else{
             return null;
-        }*/
-        return null;
+        }
     }
-    /*----------------------------------先他媽別動------------------------------------*/
     public boolean addRelationship(String userId,String deviceId){
-        User user = userRepository.findById(userId).orElse(null);
-        Device device = deviceRepository.findById(deviceId).orElse(null);
-        UserDevice userDevice = new UserDevice();
-        userDevice.setDevice(device);
-        userDevice.setUser(user);
+        User user=userRepository.findById(userId).orElse(null);
+        Device device=deviceRepository.findById(deviceId).orElse(null);
+        Set<Device> devices = user.getDevice();
+        devices.add(device);
+        user.setDevice(devices);
         try{
-            userdeviceRepository.save(userDevice);
+            userRepository.save(user);
             return true;
 
         }catch(Exception e){
@@ -167,11 +149,7 @@ public class DatabaseServiceImpl implements DatabaseService {
     }
     public List<Device.Data> queryDeviceMember(String userId){
         User user=userRepository.findById(userId).orElse(null);
-        Set<UserDevice> userDevices=user.getUserDevices();
-        List<Device> devices = new ArrayList<>();
-        for(UserDevice userDevice : userDevices){
-            devices.add(userDevice.getDevice());
-        }
+        Set<Device> devices=user.getDevice();
         List<Device.Data> members = new ArrayList<>();
         for(Device device : devices){
             members.add(device.toData());
